+++ src/game/sprite.h
@@ -157,9 +157,9 @@ s32 spr_free_sprite(s32 spriteInstanceID);
 
 s32 get_npc_comp_imgfx_idx(s32, s32);
 
-void set_npc_imgfx_comp(s32, s32, ImgFXType, s32, s32, s32, s32, s32);
+void set_npc_imgfx_comp(s32, s32, ImgFXType, intptr_t, s32, s32, s32, s32);
 
-void set_npc_imgfx_all(s32 arg0, ImgFXType arg1, s32 arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6);
+void set_npc_imgfx_all(s32 arg0, ImgFXType arg1, intptr_t arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6);
 
 s32 spr_get_comp_position(s32 spriteIdx, s32 compListIdx, s32* outX, s32* outY, s32* outZ);
